pyproject.toml
[tool.poetry]
name = "rio-temsplat"
version = "0.1.0"
description = "for customizing a generated ini file"
authors = ["Jesse Schoch"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
cheetah3 = "^3.2.6.post1"
configparser = "^7.1.0"
pytest = "^8.3.4"
multidict = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
